Rename services/data/ dir to repositories.
