+++ src/app/app.component.ts
@@ -10,16 +10,16 @@ import { CameraService } from './services/camera/camera.service';
 import { CollectorService } from './services/collector/collector.service';
 import { CapacitorProvider } from './services/collector/information/capacitor-provider/capacitor-provider';
 import { WebCryptoApiProvider } from './services/collector/signature/web-crypto-api-provider/web-crypto-api-provider';
-import { CaptionRepository } from './services/data/caption/caption-repository.service';
-import { InformationRepository } from './services/data/information/information-repository.service';
-import { ProofRepository } from './services/data/proof/proof-repository.service';
-import { SignatureRepository } from './services/data/signature/signature-repository.service';
 import { LanguageService } from './services/language/language.service';
 import { NotificationService } from './services/notification/notification.service';
 import { AssetRepository } from './services/publisher/numbers-storage/data/asset/asset-repository.service';
 import { NumbersStorageApi } from './services/publisher/numbers-storage/numbers-storage-api.service';
 import { NumbersStoragePublisher } from './services/publisher/numbers-storage/numbers-storage-publisher';
 import { PublishersAlert } from './services/publisher/publishers-alert/publishers-alert.service';
+import { CaptionRepository } from './services/repositories/caption/caption-repository.service';
+import { InformationRepository } from './services/repositories/information/information-repository.service';
+import { ProofRepository } from './services/repositories/proof/proof-repository.service';
+import { SignatureRepository } from './services/repositories/signature/signature-repository.service';
 import { SerializationService } from './services/serialization/serialization.service';
 import { fromExtension } from './utils/mime-type';
