+++ src/app/pages/home/asset/asset.page.ts
@@ -11,10 +11,10 @@ import { map, pluck, switchMap, switchMapTo, tap } from 'rxjs/operators';
 import { BlockingActionService } from 'src/app/services/blocking-action/blocking-action.service';
 import { CapacitorProvider } from 'src/app/services/collector/information/capacitor-provider/capacitor-provider';
 import { ConfirmAlert } from 'src/app/services/confirm-alert/confirm-alert.service';
-import { CaptionRepository } from 'src/app/services/data/caption/caption-repository.service';
-import { InformationRepository } from 'src/app/services/data/information/information-repository.service';
-import { ProofRepository } from 'src/app/services/data/proof/proof-repository.service';
 import { AssetRepository } from 'src/app/services/publisher/numbers-storage/data/asset/asset-repository.service';
+import { CaptionRepository } from 'src/app/services/repositories/caption/caption-repository.service';
+import { InformationRepository } from 'src/app/services/repositories/information/information-repository.service';
+import { ProofRepository } from 'src/app/services/repositories/proof/proof-repository.service';
 import { isNonNullable } from 'src/app/utils/rx-operators';
 import { ContactSelectionDialogComponent, SelectedContact } from './contact-selection-dialog/contact-selection-dialog.component';
 import { Option, OptionsMenuComponent } from './options-menu/options-menu.component';
