+++ src/app/pages/home/asset/information/information.page.ts
@@ -3,11 +3,11 @@ import { ActivatedRoute } from '@angular/router';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { map, pluck, switchMap } from 'rxjs/operators';
 import { WebCryptoApiProvider } from 'src/app/services/collector/signature/web-crypto-api-provider/web-crypto-api-provider';
-import { InformationType } from 'src/app/services/data/information/information';
-import { InformationRepository } from 'src/app/services/data/information/information-repository.service';
-import { ProofRepository } from 'src/app/services/data/proof/proof-repository.service';
-import { SignatureRepository } from 'src/app/services/data/signature/signature-repository.service';
 import { AssetRepository } from 'src/app/services/publisher/numbers-storage/data/asset/asset-repository.service';
+import { InformationType } from 'src/app/services/repositories/information/information';
+import { InformationRepository } from 'src/app/services/repositories/information/information-repository.service';
+import { ProofRepository } from 'src/app/services/repositories/proof/proof-repository.service';
+import { SignatureRepository } from 'src/app/services/repositories/signature/signature-repository.service';
 import { isNonNullable } from 'src/app/utils/rx-operators';
 
 @UntilDestroy({ checkProperties: true })
