+++ src/app/services/publisher/numbers-storage/data/asset/asset-repository.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { concatMap, concatMapTo, first, map } from 'rxjs/operators';
-import { CaptionRepository } from 'src/app/services/data/caption/caption-repository.service';
-import { Information } from 'src/app/services/data/information/information';
-import { InformationRepository } from 'src/app/services/data/information/information-repository.service';
-import { Proof } from 'src/app/services/data/proof/proof';
-import { ProofRepository } from 'src/app/services/data/proof/proof-repository.service';
-import { SignatureRepository } from 'src/app/services/data/signature/signature-repository.service';
+import { CaptionRepository } from 'src/app/services/repositories/caption/caption-repository.service';
+import { Information } from 'src/app/services/repositories/information/information';
+import { InformationRepository } from 'src/app/services/repositories/information/information-repository.service';
+import { Proof } from 'src/app/services/repositories/proof/proof';
+import { ProofRepository } from 'src/app/services/repositories/proof/proof-repository.service';
+import { SignatureRepository } from 'src/app/services/repositories/signature/signature-repository.service';
 import { SerializationService } from 'src/app/services/serialization/serialization.service';
 import { blobToDataUrlWithBase64$ } from 'src/app/utils/encoding/encoding';
 import { Storage } from 'src/app/utils/storage/storage';
