+++ src/app/services/serialization/serialization.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { first, map } from 'rxjs/operators';
-import { Importance, Information, InformationType } from '../data/information/information';
-import { InformationRepository } from '../data/information/information-repository.service';
-import { Proof } from '../data/proof/proof';
+import { Importance, Information, InformationType } from '../repositories/information/information';
+import { InformationRepository } from '../repositories/information/information-repository.service';
+import { Proof } from '../repositories/proof/proof';
 
 export type EssentialInformation = Pick<Information, 'provider' | 'name' | 'value'>;
